Add code to remove dupes from /etc/hosts
Before trying to sync to Route 53, we need to make sure the local
/etc/hosts file we've parsed doesn't have any duplicates in it, since
that will make the Route53 update fail.  Duplicates are not considered
an error and the first entry for a given name will be the one used.  The
theory behind this is that it's better for updates to continue on all
host names even in the presence of this error state.