Fix deletion of non-standard TTL records
When deleting a Route 53 record, all fields specified in the rrset must
match what is already in Route 53.  However, when deleting a record we
didn't create, or if the TTL has changed, then a mismatch will occur and
the delete will fail.  To work around that, we save a copy of the rrset
when converting to a HostEntry struct, and then use that only when
issuing a delete.